Moved workbook parsing to XlsWorkbook and worksheet parsing to XlsWorksheet. Changed worksheet parsing to using yield and updated the reader implementation
