--- src/ExcelDataReader/Core/BinaryFormat/XlsWorkbookGlobals.cs
@@ -1,41 +1,0 @@
-using System.Collections.Generic;
-
-namespace ExcelDataReader.Core.BinaryFormat
-{
-    /// <summary>
-    /// Represents Globals section of workbook
-    /// </summary>
-    internal class XlsWorkbookGlobals
-    {
-        public XlsBiffInterfaceHdr InterfaceHdr { get; set; }
-
-        public XlsBiffRecord Mms { get; set; }
-
-        public XlsBiffRecord WriteAccess { get; set; }
-
-        public XlsBiffSimpleValueRecord CodePage { get; set; }
-
-        public XlsBiffRecord Dsf { get; set; }
-
-        public XlsBiffRecord Country { get; set; }
-
-        public XlsBiffSimpleValueRecord Backup { get; set; }
-
-        public List<XlsBiffRecord> Fonts { get; } = new List<XlsBiffRecord>();
-
-        public Dictionary<ushort, XlsBiffFormatString> Formats { get; } = new Dictionary<ushort, XlsBiffFormatString>();
-
-        public List<XlsBiffRecord> ExtendedFormats { get; } = new List<XlsBiffRecord>();
-
-        public List<XlsBiffRecord> Styles { get; } = new List<XlsBiffRecord>();
-
-        public List<XlsBiffBoundSheet> Sheets { get; } = new List<XlsBiffBoundSheet>();
-
-        /// <summary>
-        /// Gets or sets the Shared String Table of workbook
-        /// </summary>
-        public XlsBiffSST SST { get; set; }
-
-        public XlsBiffRecord ExtSST { get; set; }
-    }
-}
